Remove v2 reference data registry
