+++ .github/scripts/test_alembic_v2_lifecycle.sh
@@ -4,5 +4,4 @@ set -euo pipefail
 
 uv run pytest -q \
   tests/unit/v2/test_alembic_v2.py \
-  tests/unit/v2/test_reference_data_autogenerate.py \
   tests/integration/test_alembic_v2_lifecycle.py
