+++ docs/engineering/skills/alembic-migrations.md
@@ -36,10 +36,9 @@ those narrow review corrections, stop and request a human migration decision.
 Before committing a migration, run these checks against the matching domain:
 
 1. Run `uv run alembic -c <configuration> check` and review the generated
-   schema and declared-data operations.
+   schema operations.
 2. Upgrade a fresh database to `head`.
-3. Compare the database to the complete ORM metadata and, for v2, the declared
-   application-data source before release.
+3. Compare the database to the complete ORM metadata before release.
 4. Downgrade one revision and upgrade to `head` again in an isolated database.
 5. Confirm application startup performs no implicit DDL.
 
@@ -110,42 +109,42 @@ the reviewed inventory and fail before generation or execution if expected
 tables are absent or v1, predecessor, `runtime_bundles`, population, or other
 unreviewed tables are registered.
 
-## Generated v2 application-data migrations
+## V2 application-data migrations
 
-Alembic is also the sole authority for versioned v2 application-data changes.
-Small reference data belongs in a versioned declarative source with stable
-natural identifiers and deterministic before-and-after values. The bounded v2
-autogeneration comparator and renderer compare that declaration with the
-current target and emit ordered, reversible operations through the same
-command used for schema revisions:
+Alembic is also the sole authority for intentional migrations of retained v2
+application data. Every such revision must originate from the same mandatory
+autogeneration command used for schema revisions:
 
 ```bash
 uv run alembic -c alembic-v2.ini revision --autogenerate -m "<description>"
 ```
 
 Do not create a blank revision and add `bulk_insert`, SQL strings, ORM calls,
-or other data operations by hand. Correct the declaration or generator and
-regenerate when output is wrong. Generated data additions run only after their
-required schema exists, and generated removals run before destructive schema
-operations. An unsafe identifier, unknown prior value, non-deterministic
-ordering, or non-reversible change must fail generation and invoke the human
-decision rule above.
+or other data operations by hand. A concrete data-migration requirement must
+provide a reviewed generation source capable of producing deterministic and
+reversible operations. If safe autogeneration is unavailable, invoke the human
+decision rule above instead of maintaining speculative migration machinery.
 
-`alembic check` for v2 must detect both schema drift and declared-data drift.
+Stage 8 has no active desired-state reference-data declaration or row
+comparator. Historical revisions `6ee725e0c563` and `8b8ee7fe26bb` contain the
+autogenerated insertion and removal of two synthetic validation rows. Their
+minimal operation executor remains importable only so immutable migration
+history can be replayed; the current v2 head contains neither validation row.
+
+`alembic check` for v2 must detect schema drift.
 Application startup, model import, project provisioning, and Supabase Storage
 bootstrap must never call `create_all`, create or stamp application tables, or
 mutate versioned application data. The Supabase CLI is not an application
 schema or seed migration authority.
 
-Large canonical metadata catalogs derived from the exact installed country and
+Canonical metadata catalogs derived from the exact installed country and
 `policyengine` packages are not hand-authored migration data. A later-stage,
 explicit deployment seeder may materialize those package-derived rows after
 `alembic upgrade head` when its reviewed contract requires transactional,
 idempotent row-only behavior, recorded source package versions, and fail-closed
 handling of partial catalogs. Such a seeder must perform no DDL, must not run at
 application startup, and must not transform or delete retained domain data;
-those operations remain Alembic migrations. Small reviewed reference rows stay
-in the declarative autogeneration workflow above.
+those operations remain Alembic migrations.
 
 Migration credentials remain separate from future runtime credentials. The
 migration identity may create and alter the v2 application schema; the runtime
